Pull some stuff from R3z to make robust
Here's where we are at.  I was running some tests and seeing intermittent
failures - my favorite kind.  Including some print statements to help debug
made the problem go away entirely, which clued me into the situation - we
were facing a race condition.

As it happens, this one has to do with how sockets work.  The client side
connects and can be used before the server side has finished its "accept"
phase.  Yay.  Weird.

In the process of figuring this out, I brought in a data type that can handle
multiple threads (SimpleConcurrentSet) and the ActionQueue + Logger, so that I
can log without causing such delay as to hide the problem.

With that in hand, I was able to observe what was happening, as I alluded to
before - that the server socket wasn't fully ready by the time we started
looking for it in the list of server sockets.  To fix this, I added a
little loop that will try to sleep a bit (up to 3 times) and check again.

Since this is only a problem when I'm testing, it's not the worst thing in
the world, and I don't really see any other way to gain absolute clarity on
when both sides of the TCP connection (that is, both sockets) are completely
ready.

I also added something I found online - an algorithm for "afterExecute" in
ExtendedExecutor, a subclass of the ThreadPoolExecutor, so that if anything
throws an exception it won't get hidden.