Logging toggling and Executor algorithm
Adjusted the logger so that I am able to turn various logs on or off.
Also, for the afterExecute method I just pulled the algorithm directly
out of the example in the docs for super.afterExecute.
Wrote a test to demonstrate an exception thrown in a thread submitted
to an ExecutorService.