create new "handler" for server socket
Added an option for the system to provide a "handler"
when starting the system.  This is just a function that takes
the SocketWrapper around the server's socket when a client
connects, and so basically you can do anything you want with it.

We can, and will, create an entire web framework on the back
of this.  We already did something similar in the r3z project.