Stepping away from a big file, back towards r3z approach
Listen.  The file system provided by an OS is too beautiful a thing to waste.  Why
not use it?  I was originally planning to do something like H2, wherein they use
a single big file to contain all the data.  Who knows what madness boils inside
that cauldron - but I've taken a look at the source for H2, it's huge.  Do I really
want to go that way?

I thought to myself, the R3z approach wasn't really that bad, except that every row
of data needed an id.  But why not use hash functions instead, then we don't actually
need an id.

so that's kind of where I'd like to head.  each row of data gets its own file.  The
database primarily resides in memory, and eventually (or occasionally?) writes its
data to disk.  The data is in a map structure.

Next steps: Pull in the database from R3z but make some improvements along the way, like
using a hash for its identity instead of a specified integer.  SHA-1 should do it.