More nuance in executor services
If we use the stock ExecutorService, we lose the stack trace entirely.  If we use our
custom ExtendedExecutor, it prints out the stack trace but there's no way for us to
access the Exception from above the call to create a new thread.  Now that we are
switching to using Callable rather than Thread (which is a Runnable), we are able to
access the Exception, but we have to get() on the Future created by the submit() call.

It's all really confusing.

Basically, the issue is that threads can be killed off but  unless we have some cool
mechanism to occasionally check on the results of kicked-off threads, we have no way to access the exceptions that might have killed them.  Perhaps some fancy program that
 loops through threads and expects them all to be lively, and if not to pull the
 exception higher - except where does it stop?  We don't want threads to necessarily
 be able to kill the entire program.

So in the meantime, to avoid this pain, if we use the extended executor then it does at
least show the stack trace in the logs.