Going down a new avenue for database design
Instead of holding to the strongly-typed object-oriented approach, I will
try to simply store the data as string in a 2 dimensional array, or in other
words, a table.  When I work with data it will be as strings, and I will have
to convert to/from when necessary.  A lot of the time though, it may not even
be necessary.  For example, if I just want to count the number of rows
with the data "foo" in a column, I can just do a search without any
casting to a type.

For now, since I am looking at a whole new approach for the database, I
am shutting off the OwnDatabaseTests file.