Trying a new tack with data
I want to keep things as simple as possible.  I like the idea of storing the data
as strings, but I do want to keep the ability to address discrete elements, and
to track their associated property/field names, and which class all this data came
from.

This test shows that approach.