various improvements
according to recent research, using the plus operator in Java is actually a really
fast way to append strings, and format is slow - switching to that in several cases.

Also, since the current agenda is to have multiple DatabaseDiskPersistenceSimpler's,
we'll name the actionqueue of each one per the data type.

adding a performance test for updating data, bypassing synchronizing to disk.

adjusting the required wait time for thread sleep in a test to the bare minimum,
by adjusting until errors occurred and then giving it a bit of wiggle room.