refactoring and redesign
Converting some classes to records (so I don't have to create a
custom toString method for them).

adding a capability to return a simplistic 404 if the client requests
an unregistered endpoint.  This was happening a lot since the browser
was asking for favicon.

Adjusted the main class to actually do something meaningful, instead of
just returning the sum of two numbers to the console.  Now, it starts
up a web server with an endpoint.  To test this out, run the application
and open a browser to http://localhost:8080/add_two_numbers?a=42&b=46

TODO:
 - the response is always a 200 ok.  Adjust the response record so it can
   be different types.