Naive persistence of sample domain
Just getting the ball rolling here...

created a simple domain concept, "PersonName", which has persistence set up.

To test this out, run the application:

    make run

Then, in the browser, go to:

   http://localhost:8080/formentry

Type in whatever and press enter, it should store that data into out/simple_db/names
with files named per the object's index (1.db, 2.db, etc), with contents like "1 abc".

This is just a proof-of-concept. Knowing that it works, we can now make strides to
build simple patterns for more correct usage.