static responses are now returned.  New Trace level for logs.
Didn't realize until today, but when we use the class loader to get some directory
in our resources, it goes through all the folders and jars in the classpath provided.

To have it properly load the data in "static/", I just needed to add the resources
directory to the classpath.  But first, I needed to copy that directory to the "out/"
just to be more consistent - wouldn't want to be grabbing files willy-nilly.

Although come to think of it, we *are* pulling in jars from lib in the root directory.
Ah well, a problem for another day.

Added a trace level in logging so we can hide away the noisier parts of our logs.