Block at a precise place
One thing I found out when trying out the virtual threads executor is
that the system would just fall through if I didn't have something to
"block" it at a certain point.  This line performs that service.  It
sits there, blocking, until it "gets" a result from the central loop,
which, oh by the way, is wrapping an infinite loop that only exits if we
interrupt it forcibly (like by the command in Server.stop and
FullSystem.shutdown, etc.)