Correcting a bug
Found that the system was locking up during tests.  Looked like it was getting caught somewhere in testFullSystem_Soup_To_Nuts, related to the try-with-resources.  Once I changed that, I began seeing the runtime exception messages that had been apparently swallowed up before. What was happening was a parsing error with deserializing PersonNames.  Once I saw that, it was easily fixed.