Further auth testing, refactoring, documentation
Noticed that when exceptions were thrown in tests, it would interleave with log statements.  Adjusted so the output of exceptions is inline with log statements.

Some repetitive approaches currently underway for handling auth in sample domain, but this is on purpose to see where we ought to abstract later on.