Hitting a wall - HTTP body is a string
Ok, looks like some of our earlier decisions surrounding the design of HTTP requests/responses
needs reconsideration.  Until this point, we have been treating the data as strings, mostly.  But
if we need to receive a request that has a binary body, that has a cascade of effects on things.

For example, when we read data on a request, we go line-by-line.  But in binary-land, that is not
a thing.  And, how can we pre-parse the body into a map of keys and values if the body might be
an image?