Make Headers slightly stronger-typed.
We may want to redesign this soon. It works for our situation, but if we want to have easy access
to flags and options from the headers, there might be a better way than the current direction.

Tests are still failing for the new photo domain, but that's primarily because it has not become
apparently necessary to add multipart to our kit, rather than solely relying on url-encoding.