Adjusting to read bytes at a time. Improve test output
When we are reading HTTP data, we have to vary our approach depending on the state of things.
More specifically, when we read the first incoming line (e.g. GET /foo), we must read up to the
CR + LF.  Similarly for each following header line.  But after those lines, when reading the body,
we need the ability to read continuous blocks of data - it might be an image or other binary.

If we use a buffered reader, this breaks things for us, because it is the nature of the buffered
reader to "read ahead", and we don't want to do that in this protocol.  So now, these code changes
will cause the reader to read one byte at a time, so we are at precisely the right point in our reading.

This will empower us to read images and other binaries as part of an upcoming feature.

Noticed room for an improvement in our testing framework.  When a string has whitespace, it
now shows the whitespace as (RETURN), (TAB), etc. so you can more apparently see the differences.

Newest test "A user should be able to send a photo" is still broken - not yet implemented.

Removed some redundant code.