Basic photo domain working, with some fake stuff
All tests passing

had to ugly it up a bit, but the repetitiveness is intentional and begs for refactoring.

Sending and receiving data by HTTP are now better aligned with how it really happens,
which is by bytes (rather than Strings).  This leads us to use Map<String, Object> for the
body map, and we have to cast data types, ... like I said, it's ugly.  But,
a) it works
b) it's ready for polishing once we figure out what is essential.