improving broken test, adding new broken test
Some research indicated that when a client does a POST to our server,
it absolutely must have either a content-length header or do it
with transfer-encoding: chunked.  Previously, the existing test had
neither of these headers, which would make it invalid.  Adjusted the
existing test to have a content-length header.

Also, added a new test following it to drive creation of code for
handling chunked data.