Massive renaming to Minum
Reason: The old name had no connection to where I want to go.  The old name,
ATQA, stood for Automated Test Quality Analyzer, because what started this
was trying to create a tool to assess test quality of a project.  As part
of that I was creating some tangential code in a high-quality fashion, in
order to set a high bar touchstone for the tool.  Then I found out that
writing that tool was going to be pretty complicated and almost definitely
require using dependencies, since I would effectively need the ability to
compile Java code to instrument it properly.  Ugh.

So instead I kept building out the web framework, and the term ATQA no longer
fit its intended direction.  So I chose "Minum", relating to terms like
minimalism and minimum viable product, which were overarching paradigms here.