Refactoring constants and config, adding TheBrig
To avoid some boilerplate, and to make things more consistent, all the
configurable options are now in both the app.config file and the Constants
file.  This way, anything we configured has documentation that should show up
in the IDE.

Added TheBrig, a security tool for putting potential attackers into a timed
queue, where (if enabled), we will drop their connection for a duration.