Improving some test coverage, removing unused test logger
One of the test loggers was especially designed to collect
logging information for use in assertions.  Now, I just have
a queue in the regular test logger that collects the
last 30 log entries.  Note that this test logger has
synchronized methods, so it will be slower than the normal
logger.