Switch to lazy loading for static files
Instead of looping through all the files in the static directory and adding
them to the cache, we'll let user requests cause us to load them.  The benefit
is that the startup becomes marginally faster if we don't need to load
files as part of startup.  Admittedly, this incurs an extra cost the first time
each file is loaded, but the benefit is that when we need to send up new
binaries, they are up and running that much sooner.