Db data structure
The data structure used for data in the database has switched
from a list to a map.

The main reason is that we want faster access to elements by
their index value.

During delete and update operations, previously we needed to
run a linear search for an element by index - an O(n) operation.
With this improvement, it becomes O(1).

This also gives us some extra fancy operations, like how instead
of exposing a .stream() interface to the list, we can expose a
Collections, which enables far greater versatility with the data.