Handle endpoint error better
Before this change, if an exception was thrown in an endpoint it would bubble up
to the top of the server and get logged as an ASYNC_ERROR, and the server would
respond by just closing the socket.  Gross!

Now, a bit deeper down and closer to the endpoint, we wrap the call with a try-catch,
and if anything fails in the endpoint we return a 500 to the client, with a randomly
generated code that is put on both the client response message and the log with
more detail.