More sophistication in filesystem use
This part of our code is used when we need to access data at runtime that is stored
in our (compressed) jar file.  Not 100% sure but I think it has to create a file
system to temporarily decompress the data, and then maybe it evaporates right after?

Anyway, I was finding a weird edge case when I would run the system on the Oracle
Linux box - it would complain that a file system already existed.  So I wrote this
to initially look for the file system that exists, and if that fails fall
back to creating a new one.

I didn't see a command to check if a filesystem exists without it throwing an
exception.

Yes, there are performance problems when relying on exception handling as a normal
part of control flow, but

a) I couldn't find any other way to look for an existing system
b) this only runs when looking for static assets, which are finite and few,
   that are stored in our jar file, and once they are obtained they are
   served from an in-memory cache instead of the disk.

Unfortunately this is one of those areas that rely on implementation from the
OS vendors and can thus have widely-varying outcomes.  I don't feel like I have
a good grasp on this.