Improve error handling
If the response object gets to the rendering stage with a body but no content type,
the system throws an error.

This is to maintain correctness of what it is producing.  In fact, the browser has some
smarts about figuring things out, some of the time.  But if we want to maintain
a valid system, any body must be tagged with its content-type mime.

Previous to this change, an error message would mention this and also indicate the headers
on the response object.  Now, we include more detail on the response, as well as many
parts of the request object too, to better situate a developer.