Locking down the system - final on classes
Got a comment online about preventing inheritance on classes not intended to be
inherited.  Hadn't really considered the drawbacks of that - read through Josh
Bloch's Effective Java and thought about the recommendations.  Still seems a bit
unlikely - which concrete classes would people extend? Why?

There's a part of me that wants to allow it to remain inheritable, because I want
to allow flexibility.  But maybe they're right - lock it down, and if I hear or
find arguments otherwise, I can always open it up later.  But if people start
relying on inheritance when I hadn't designed for it, maybe a problem?  Ah well.

All the tests ran just fine.