Massive redesign - static file cache
The advantage of this change is that the static files will sit outside the binary, so
it becomes possible to adjust them while the system is running and immediately see
the results, without needing to restart the system.  It also simplifies the system a
bit and lets us remove some code.

We no longer need to create a temporary file system when we decompress files from
our jar.

The only file left in the jar is the keystore, which is used for certs when testing
on localhost.

This change causes a need for a new item in the app.config, KEEP_ALIVE_TIMEOUT_SECONDS