v2.1.2 release
* Address a bug in the templating code
* Apply suggestions from Sonarqube
* Update documentation

There was a bug in the templating code.  It turned out that there was an edge
case around closing double-curly-braces.  The parser would presume, upon
encountering the closing braces, that the prior text had been a template
literal.  This was a bug, and a test-driven correction was applied.

Also, ran Sonarqube and applied a variety of improvements it suggested.