v2.4.0 - Documentation for ActionQueue.enqueue, assertion failure messages
The behavior of ActionQueue was adjusted - it takes ThrowingRunnable for an
action instead of Callable, which means it is not necessary to return null.
The documentation and example was out of date.

Now, all the assertion methods take a failure message string to be shown
in case of a test failure.