handle bad URL-encodings better - simply skip
It was discovered that if a user sent a request with improperly-formed URL-encoded data, it would
sometimes cause a 500 error.  This was not a significant issue, but ergonomically the system should
not react so strongly for a relatively minor issue.

This adjustment takes a milder approach, simply moving on to the next key-value pair and logging
about the action.