v8.3.2 - handle unread bodies. Small various improvements.
Handling unread bodies
----------------------

if the user is doing unexpected things like sending a request to a POST handler
with a body but then ignoring the body, a safe way to handle it is to just
close the socket.  The browser is perfectly capable of handling a closed socket
- it will just send the next request on a new socket.

The problem we are trying to resolve is when the server is dealing with
something like an attacker who is sending requests with large bodies and the
system configured to allow them.  If we just slurp up those bytes, we end up
letting a lot of bytes through our network connection for nothing.

This is a conservative and security-oriented approach for an unusual situation.

Handling null better in assertions
----------------------------------

Some of the assertion code, like assertEquals, does not work properly if one
of the inputs is a null.  Code was adjusted so any null inputs will cause a clear
exception to be thrown, advising user to an alternate approach like assertTrue(foo == null)

Greater consistency in handling duplicate keys in query strings and bodies
--------------------------------------------------------------------------

Here is an example of a duplicate key, "foo", in a query string:

    GET http://localhost/endpoint?foo=baz&foo=bar

In that case, we will add a log about how the last-one-in wins.  We will do the same when
a URL-encoded body includes duplicate keys.